+++ deploy/15-deploy-lbtc-oracle.ts
@@ -1,0 +1,32 @@
+import hre from "hardhat";
+import { DeployFunction } from "hardhat-deploy/dist/types";
+import { HardhatRuntimeEnvironment } from "hardhat/types";
+
+import { ADDRESSES } from "../helpers/deploymentConfig";
+
+const func: DeployFunction = async function ({ getNamedAccounts, deployments, network }: HardhatRuntimeEnvironment) {
+  const { deploy } = deployments;
+  const { deployer } = await getNamedAccounts();
+  const proxyOwnerAddress = ADDRESSES[network.name].timelock;
+  const { LBTC, WBTC } = ADDRESSES[network.name];
+
+  const redstoneOracle = await hre.ethers.getContract("RedStoneOracle");
+  const resilientOracle = await hre.ethers.getContract("ResilientOracle");
+
+  await deploy("LBTCOneJumpRedStoneOracle", {
+    contract: "OneJumpOracle",
+    from: deployer,
+    log: true,
+    deterministicDeployment: false,
+    args: [LBTC, WBTC, resilientOracle.address, redstoneOracle.address],
+    proxy: {
+      owner: proxyOwnerAddress,
+      proxyContract: "OptimizedTransparentProxy",
+    },
+    skipIfAlreadyDeployed: true,
+  });
+};
+
+func.skip = async () => hre.network.name !== "ethereum" && hre.network.name !== "sepolia";
+func.tags = ["LBTCOneJumpRedStoneOracle"];
+export default func;